QHDEV-734 - Move the animation engine to src/utils/animate.js
`animate` is a generic engine — it steps a CSS property from its current
value to a target on any element and knows nothing about our markup, our
class names or any component. Living under `src/components/_global/js/`
implied it was a component, and gave it a two-segment path
(`animate/global.js`) for a single module.

It is now `src/utils/animate.js`, alongside the other helpers with no
component knowledge.

The module body is unchanged, including the PascalCase export names:
those are the `QLD.animate.*` legacy contract for external callers, so
the docblock now records why they do not match the rest of `src/utils`.
