+++ src/components/_global/js/legacyGlobal.js
@@ -18,7 +18,7 @@
 // exports; the object is then spelled out member-by-member below so the legacy
 // contract stays explicit and a removed source export surfaces here.
 import utils from "./global.js";
-import * as animate from "./animate/global.js";
+import * as animate from "../../../utils/animate.js";
 import { tabs } from "./tabs/global.js";
 import { Modal } from "./modal/global.js";
 import { toggleToolTips } from "./popover/controller.js";
